% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{add_all_cogs}
\alias{add_all_cogs}
\title{Add a column to a data frame containing cognostics for fitted trajectories and set longitudinal cognostics}
\usage{
add_all_cogs(dat, x_units = "days", recovery = NULL)
}
\arguments{
\item{dat}{a data frame nested by subject and containing fits from \code{\link{fit_all_trajectories}}}

\item{x_units}{units of age variable ("days", "months", or "years")}

\item{recovery}{age in days at which to plot recovery from nadir}
}
\description{
Add a column to a data frame containing cognostics for fitted trajectories and set longitudinal cognostics
}

