% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{add_longi_cogs}
\alias{add_longi_cogs}
\title{Set longitudinal cognostics for subject-nested data frames}
\usage{
add_longi_cogs(dat)
}
\arguments{
\item{dat}{a data frame nested by subject and containing fits from}
}
\description{
Set longitudinal cognostics for subject-nested data frames
}
