% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{add_trajectory_plot}
\alias{add_trajectory_plot}
\alias{add_velocity_plot}
\title{Add a column to a data frame containing trajectory / velocity plots}
\usage{
add_trajectory_plot(dat, z = FALSE, center = FALSE, y_var = "htcm",
  x_range = NULL, nadir = FALSE, recovery = NULL, x_units = "days",
  width = 500, height = 520)

add_velocity_plot(dat, z = FALSE, center = FALSE, x_range = NULL,
  nadir = FALSE, recovery = NULL, x_units = "days", width = 500,
  height = 520)
}
\arguments{
\item{dat}{a data frame nested by subject and containing fits from \code{\link{fit_all_trajectories}}}

\item{z}{should the trajectory fit be plotted on the z-scale?}

\item{center}{should the trajectory be centered around the median WHO standard?}

\item{y_var}{name of y variable to model (usually an anthropometric measure or z-score scaled anthropometric measure)}

\item{x_range}{a vector specifying the range (min, max) that the superposed WHO growth standard should span on the x-axis}

\item{nadir}{should a guide be added to the plot showing the location of the nadir? (only valid when z = TRUE)}

\item{recovery}{age in days at which to plot recovery from nadir (only valid when z = TRUE) - if NULL (default), will not be plotted}

\item{x_units}{units of age x-axis ("days", "months", or "years")}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\description{
Add a column to a data frame containing trajectory / velocity plots
}
\examples{
\dontrun{
cppsubj  <- by_subject(cpp)
cppfit   <- get_fit(cpp, method = "rlm")
cpptr    <- fit_all_trajectories(cppsubj, cppfit)
cppplot  <- trscope_trajectories(cpptr)
cppzplot <- trscope_trajectories(cpptr, z = TRUE, nadir = TRUE, x_units = "months")
}
}

