% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{auto_loess}
\alias{auto_loess}
\title{Find best loess fit based on aic or gcv}
\usage{
auto_loess(data, span = c(0.01, 2), degree = c(1, 2), family = "gaussian",
  which = "gcv", ...)
}
\arguments{
\item{data}{a data frame with columns x and y to be used in the fitting}

\item{span}{a vector indicating a range (min, max) of spans to search over}

\item{degree}{a vector of degrees to search over (valid values are combinations of 0, 1, 2)}

\item{family}{loess \code{family} parameter}

\item{which}{which method to use, "aicc" or "gcv"}

\item{\ldots}{additional parameters passed to \code{\link{loess}}}
}
\description{
Find best loess fit based on aic or gcv
}

