% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions.R
\name{by_trajectory_checkpoints}
\alias{by_trajectory_checkpoints}
\title{Split by-subject trajectory-fitted data by checkpoint categorizations}
\usage{
by_trajectory_checkpoints(dat, complete = TRUE)
}
\arguments{
\item{dat}{a data object returned by \code{\link{fit_all_trajectories}}}

\item{complete}{subset only to those that have fitted checkpoints}
}
\description{
Split by-subject trajectory-fitted data by checkpoint categorizations
}
\examples{
\dontrun{
cppsubj <- by_subject(cpp)
cppfit  <- get_fit(cpp, method = "rlm")
cpptr   <- fit_all_trajectories(cppsubj, cppfit)
cppcp   <- by_trajectory_checkpoints(cpptr)
}
}

