% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check a dataset to ensure it will be compatible with hbgd methods}
\usage{
check_data(dat, has_height = TRUE, has_weight = TRUE, has_hcir = TRUE)
}
\arguments{
\item{dat}{a data frame}

\item{has_height}{does this dataset contain anthropometric height data?}

\item{has_weight}{does this dataset contain anthropometric weight data?}

\item{has_hcir}{does this dataset contain anthropometric head circumference data?}
}
\description{
Check a dataset to ensure it will be compatible with hbgd methods
}
\examples{
check_data(cpp, has_hcir = FALSE)

smc <- brokenstick::smocc.hgtwgt
check_data(smc, has_hcir = FALSE)

names(smc)[2] <- "subjid"
names(smc)[5] <- "agedays"
smc$sex <- as.character(smc$sex)
smc$sex[smc$sex == "male"] <- "Male"
smc$sex[smc$sex == "female"] <- "Female"
names(smc)[10] <- "htcm"
names(smc)[11] <- "wtkg"

check_data(smc, has_hcir = FALSE)

names(smc)[12] <- "haz"
smc$waz <- who_wtkg2zscore(smc$agedays, smc$wtkg, smc$sex)
smc$agedays <- smc$agedays * 365.25

check_data(smc, has_hcir = FALSE)
}

