% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.brokenstick}
\alias{fit_method.brokenstick}
\title{Get the result of fitting brokenstick to a dataset}
\usage{
\method{fit_method}{brokenstick}(dat, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{\ldots}{additional parameters passed to \code{\link[brokenstick]{brokenstick}}, most notably \code{knots}}
}
\description{
Get the result of fitting brokenstick to a dataset
}
\details{
This essentially gets an anthropometric dataset into shape for \code{\link[brokenstick]{brokenstick}} (sets appropriate data structure and removes missing values) and runs the fitting routine.
}
\note{
The settings for \code{x_trans} and \code{y_trans} must match that used in \code{\link{fit_trajectory}} and appropriate inverse transformations must be set there accordingly as well.
}
\examples{
\dontrun{
bsfit <- get_fit(cpp, y_var = "haz", method = "brokenstick")
fit <- fit_trajectory(subset(cpp, subjid == 2), fit = bsfit)
plot(fit)
}
}

