% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.face}
\alias{fit_method.face}
\title{Compute fpca "face" fit of growth trajectory}
\usage{
\method{fit_method}{face}(dat, xg = NULL, cpx = NULL, fit)
}
\arguments{
\item{dat}{data frame specifying x and y coordinates of data to fit}

\item{xg}{x grid across which to compute fits for plotting (default is sequence along the range of \code{x})}

\item{cpx}{x values at which to compute "checkpoints" of the fitted subjects's growth trajectory to compare to other subjects}

\item{fit}{the result of running \code{\link[face]{face.sparse}} against the entire data set (with convenience wrapper \code{\link{get_face_fit}}) which is used to compute per-subject fits}
}
\description{
Compute fpca "face" fit of growth trajectory
}
\note{
The trajectory fitting functions are most easily accessed through calling \code{\link{fit_trajectory}} with the \code{method} argument to specify the modeling approach to use.

These fitting functions can easily be replaced by simply calling the associated R methods, but are provided for convenience to standardize input/output to simplify swapping fitting methods.
}

