% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.loess}
\alias{fit_method.loess}
\title{Compute loess fit of growth trajectory}
\usage{
\method{fit_method}{loess}(dat, xg = NULL, cpx = NULL, span = c(0.05, 3),
  degree = c(1, 2), family = "symmetric", ...)
}
\arguments{
\item{dat}{data frame specifying x and y coordinates of data to fit}

\item{xg}{x grid across which to compute fits for plotting (default is sequence along the range of \code{x})}

\item{cpx}{x values at which to compute "checkpoints" of the fitted subjects's growth trajectory to compare to other subjects}

\item{span}{\code{\link{loess}} smoothing parameter}

\item{degree}{\code{\link{loess}} smoothing parameter}

\item{family}{\code{\link{loess}} smoothing parameter}

\item{\ldots}{additional parameters passed to \code{\link{loess}}}
}
\description{
Compute loess fit of growth trajectory
}
\note{
The trajectory fitting functions are most easily accessed through calling \code{\link{fit_trajectory}} with the \code{method} argument to specify the modeling approach to use.

These fitting functions can easily be replaced by simply calling the associated R methods, but are provided for convenience to standardize input/output to simplify swapping fitting methods.
}

