% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.lwmod}
\alias{fit_method.lwmod}
\title{Get the result of fitting a Laird and Ware linear or quadratic model to a dataset}
\usage{
\method{fit_method}{lwmod}(dat, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{\ldots}{additional parameters, most notably \code{deg} which controls the degree of polynomial for the fit (1 for linear and 2 for quadratic)}
}
\description{
Get the result of fitting a Laird and Ware linear or quadratic model to a dataset
}
\details{
This essentially gets an anthropometric dataset into shape for \code{\link[sitar]{sitar}} (sets appropriate data structure and removes missing values) and runs the fitting routine.
}
\note{
The settings for \code{x_trans} and \code{y_trans} must match that used in \code{\link{fit_trajectory}} and appropriate inverse transformations must be set there accordingly as well.
}
\examples{
\dontrun{
lwfit <- get_fit(cpp, y_var = "haz", method = "lwmod", deg = 2)
fit <- fit_trajectory(subset(cpp, subjid == 2), fit = lwfit)
plot(fit)
}
}

