% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.smooth.spline}
\alias{fit_method.smooth.spline}
\title{Compute smooth.spline fit of growth trajectory}
\usage{
\method{fit_method}{smooth.spline}(dat, ...)
}
\arguments{
\item{dat}{data frame specifying x and y coordinates of data to fit}

\item{\ldots}{additional parameters passed to \code{\link[mgcv]{gam}}}
}
\description{
Compute smooth.spline fit of growth trajectory
}
\note{
The trajectory fitting functions are most easily accessed through calling \code{\link{fit_trajectory}} with the \code{method} argument to specify the modeling approach to use.

These fitting functions can easily be replaced by simply calling the associated R methods, but are provided for convenience to standardize input/output to simplify swapping fitting methods.
}

