% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysis_fit_traj.R
\name{fit_trajectory}
\alias{fit_trajectory}
\title{Fit a model to an individual's trajectory and optionally add checkpoints}
\usage{
fit_trajectory(dat, x_var = "agedays", y_var = "htcm", method = "gam",
  xg = NULL, checkpoints = 365 * c(1:2), z_bins = c(-2, 2),
  x_trans = NULL, x_inv = NULL, y_trans = NULL, y_inv = NULL, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{x_var}{name of x variable to model}

\item{y_var}{name of y variable to model}

\item{method}{one of "gam", "loess", "fda", "rlm"}

\item{xg}{grid of x points at which the fit should be evaluated for plotting (if \code{NULL} it will be set to an equally-spaced grid of 150 points across \code{x})}

\item{checkpoints}{x values at which to compute "checkpoints" of the subjects's growth trajectory to compare to other subjects}

\item{z_bins}{a vector indicating binning of z-scores for the subject's trajectory at each checkpoint with respect to the the WHO growth standard}

\item{x_trans,y_trans}{transformation functions to be applied to x and y prior to modeling}

\item{x_inv,y_inv}{inverse transformation functions for x and y to get back to the original scale after modeling}

\item{\ldots}{parameters passed on to the fitting method}
}
\description{
Fit a model to an individual's trajectory and optionally add checkpoints
}
\examples{
fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
plot(fit$xy$x, fit$xy$y)
lines(fit$fitgrid$x, fit$fitgrid$y)
# there is also a plot method:
plot(fit, who_range = c(0, 2560))
# we can fit the z-scores instead
fit2 <- fit_trajectory(subset(cpp, subjid == 2), y_var = "waz", method = "rlm")
plot(fit2$xy$x, fit2$xy$z)
lines(fit2$fitgrid$x, fit2$fitgrid$z)
}

