% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{get_agefreq}
\alias{get_agefreq}
\title{Get age frequency}
\usage{
get_agefreq(dat, agevar = "agedays", age_range = NULL)
}
\arguments{
\item{dat}{a longitudinal growth study data set}

\item{agevar}{variable name in \code{dat} that contains a measure of the subject's age}

\item{age_range}{optional range to ....}
}
\description{
Get age frequency
}
\examples{
\donttest{
agefreq <- get_agefreq(cpp)
plot_agefreq(agefreq)
}
}

