% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_avail_methods}
\alias{get_avail_methods}
\title{Get names of all available fitting methods}
\usage{
get_avail_methods()
}
\description{
Get names of all available fitting methods
}
\seealso{
\code{\link{get_fit}}
}

