% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{get_face_fit}
\alias{get_face_fit}
\title{Get the result of fitting face.sparse to a data set}
\usage{
get_face_fit(dat, x_var = "agedays", y_var = "htcm", knots = 10,
  x_trans = NULL, y_trans = NULL, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{x_var}{name of x variable to model (default "agedays")}

\item{y_var}{name of y variable to model (default "htcm")}

\item{knots}{number of knots, sent to \code{\link[face]{select_knots}}}

\item{x_trans, y_trans}{transformation functions to be applied to x and y prior to modeling (see note)}

\item{\ldots}{additional parameters passed to \code{\link[face]{face.sparse}}}
}
\description{
Get the result of fitting face.sparse to a data set
}
\details{
This essentially gets an anthropometric data set into shape for \code{\link[face]{face.sparse}} (sets appropriate data structure and removes missing values) and runs the fitting routine.
}
\note{
The settings for \code{x_trans} and \code{y_trans} must match that used in \code{\link{fit_trajectory}} and appropriate inverse transformations must be set there accordingly as well.
}
\examples{
\donttest{
facefit <- get_face_fit(cpp, y_var = "haz")
fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "haz",
  method = "face", fit = facefit)
plot(fit)
}
}

