% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_get_fit.R
\name{get_fit}
\alias{get_fit}
\title{Obtain a trajectory "fit" object for a dataset}
\usage{
get_fit(dat, x_var = "agedays", y_var = "htcm", method = "fda",
  holdout = FALSE, x_trans = NULL, x_inv = NULL, y_trans = NULL,
  y_inv = NULL, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{x_var}{name of x variable to model (default is "agedays")}

\item{y_var}{name of y variable to model (usually an anthropometric measure or z-score scaled anthropometric measure)}

\item{method}{name of fitting method to use (see \code{\link{get_avail_methods}})}

\item{holdout}{should an observation be held out for fitting (will use column \code{hold} in \code{dat} to which observations to hold out)}

\item{x_trans, y_trans}{transformation functions to be applied to x and y prior to modeling}

\item{x_inv, y_inv}{inverse transformation functions for x and y to get back to the original scale after modeling}

\item{\ldots}{parameters passed on to the fitting method}
}
\description{
Obtain a trajectory "fit" object for a dataset
}

