% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_holdout.R
\name{get_fit_holdout_errors}
\alias{get_fit_holdout_errors}
\title{Get holdout errors}
\usage{
get_fit_holdout_errors(d, z = TRUE)
}
\arguments{
\item{d}{an object returned from \code{\link{fit_all_trajectories}}}

\item{z}{compute MSE on z-score scale or original scale?}
}
\description{
Get holdout errors
}

