% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/analysis_holdout.R
\name{get_fit_holdout_mse}
\alias{get_fit_holdout_mse}
\title{Get MSE for holdout}
\usage{
get_fit_holdout_mse(d, z = TRUE)
}
\arguments{
\item{d}{an object returned from \code{\link{fit_all_trajectories}}}

\item{z}{compute MSE on z-score scale or original scale?}
}
\description{
Get MSE for holdout
}

