% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{get_recovery}
\alias{get_recovery}
\title{Get recovery statistics of z-scale growth trajectory}
\usage{
get_recovery(obj, nadir = NULL, at = 365.25 * 3)
}
\arguments{
\item{obj}{object created from \code{\link{fit_trajectory}}}

\item{nadir}{object created from \code{\link{get_nadir}} (if NULL, will be automatically generated)}

\item{at}{age (in days) at which to estimate recovery}
}
\description{
Get recovery statistics of z-scale growth trajectory
}
