% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{get_smocc_data}
\alias{get_smocc_data}
\title{Get SMOCC data from brokenstick, transformed to be hbgd-compatible}
\usage{
get_smocc_data()
}
\description{
Get SMOCC data from brokenstick, transformed to be hbgd-compatible
}

