% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_avg_vel.R
\name{get_traj_velocity}
\alias{get_traj_velocity}
\title{Returns the veclocity of growth trajectories for _all_ subjects}
\usage{
get_traj_velocity(all_traj, z_score = TRUE, avg_over_days = 365L,
  FUN = mean_na_rm_generator())
}
\arguments{
\item{all_traj}{Trajectory object returned from \code{\link{fit_all_trajectories}}.
Assumes all required data are present.}

\item{z_score}{Boolean indicating whether 'y_var' parameter was for
z-score data or not. Default is TRUE.}

\item{avg_over_days}{Number of days to averge the derivative (aka velocity)
Default is 365.}

\item{FUN}{A function that will be applied to the first derivative of
the fitted growth trajectory. This function should accept a single
argument; a vector containing the values of the first derivative.
Default value is a _{\link[curried]{https://en.wikipedia.org/wiki/Currying}}_
version of the \code{\link{mean}} function, which removes
\code{NA} values by default.}
}
\value{
A \code{\link[tibble]{tibble}} with two columns, \code{subjid}
        and \code{avg_vel}
}
\description{
Returns the veclocity of growth trajectories for _all_ subjects
}
