% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{get_x_range}
\alias{get_x_range}
\title{Get the x-axis range across all subjects}
\usage{
get_x_range(dat, pad = 0.07)
}
\arguments{
\item{dat}{object obtained from \code{\link{fit_all_trajectories}}}

\item{pad}{padding factor - this much space as a multiple of the span of the x range will be added to the min and max}
}
\description{
Get the x-axis range across all subjects
}

