% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/hbgd-package.R
\docType{package}
\name{hbgd-package}
\alias{hbgd}
\alias{hbgd-package}
\title{hbgd: Healthy Birth, Growth & Development}
\description{
\url{http://hafen.github.io/docs-hbgd/}
}
\examples{
help(package = hbgd)
}

