% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/hbgd-package.R
\docType{data}
\name{hbgd_labels_df}
\alias{hbgd_labels_df}
\title{Labels for common variable names in hbgd data}
\usage{
hbgd_labels_df
}
\description{
Labels for common variable names in hbgd data, used in \code{\link{get_data_attributes}} if labels are not explicitly provided.
}
\keyword{data}

