% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_igb.R
\name{igb_value2centile}
\alias{igb_value2centile}
\alias{igb_value2zscore}
\title{Convert birth measurements to INTERGROWTH z-scores/centiles (generic)}
\usage{
igb_value2centile(gagebrth, val, var = "lencm", sex = "Female")

igb_value2zscore(gagebrth, val, var = "lencm", sex = "Female")
}
\arguments{
\item{gagebrth}{gestational age at birth in days}

\item{val}{the value(s) of the anthro measurement to convert}

\item{var}{the name of the measurement to convert ("lencm", "wtkg", "hcircm")}

\item{sex}{"Male" or "Female"}
}
\description{
Convert birth measurements to INTERGROWTH z-scores/centiles (generic)
}
\examples{
# get Male birth length z-scores
# first we need just 1 record per subject with subject-level data
cppsubj <- get_subject_data(cpp)
cppsubj <- subset(cppsubj, sex == "Male")
igb_value2zscore(cpp$gagebrth, cpp$birthlen, var = "lencm", sex = "Male")
}
\references{
International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project
Villar, José et al.
The Lancet, Volume 384, Issue 9946, 857-868
}

