% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_birth.R
\name{igb_zscore2lencm}
\alias{igb_centile2hcircm}
\alias{igb_centile2lencm}
\alias{igb_centile2wtkg}
\alias{igb_zscore2hcircm}
\alias{igb_zscore2lencm}
\alias{igb_zscore2wtkg}
\title{Convert Intergrowth z-scores/centiles to prenatal ultrasound measurements}
\usage{
igb_zscore2lencm(gagebrth, z = 0, sex = "Female")

igb_zscore2wtkg(gagebrth, z = 0, sex = "Female")

igb_zscore2hcircm(gagebrth, z = 0, sex = "Female")

igb_centile2lencm(gagebrth, p = 50, sex = "Female")

igb_centile2wtkg(gagebrth, p = 50, sex = "Female")

igb_centile2hcircm(gagebrth, p = 50, sex = "Female")
}
\arguments{
\item{gagebrth}{gestational age at birth in days}

\item{z}{z-score(s) to convert}

\item{sex}{"Male" or "Female"}

\item{p}{centile(s) to convert (must be between 0 and 100)}
}
\description{
Convert Intergrowth z-scores/centiles to prenatal ultrasound measurements
}
\examples{
# get 99th centile for Male birth weights across some gestational ages
igb_centile2wtkg(232:300, 99, sex = "Male")
}
\references{
International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project
Villar, José et al.
The Lancet, Volume 384, Issue 9946, 857-868
}

