% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_pre.R
\name{igpre_value2zscore}
\alias{igpre_value2centile}
\alias{igpre_value2zscore}
\title{Convert prenatal ultrasound measurements to Intergrowth z-scores/centiles (generic)}
\usage{
igpre_value2zscore(gagedays, val, var = c("hccm", "bpdcm", "ofdcm", "accm",
  "flcm"))

igpre_value2centile(gagedays, val, var = c("hccm", "bpdcm", "ofdcm", "accm",
  "flcm"))
}
\arguments{
\item{gagedays}{gestational age in days}

\item{val}{the value(s) of the anthro measurement to convert}

\item{var}{the name of the measurement to convert ("hccm", "bpdcm", "ofdcm", "accm", "flcm")}
}
\description{
Convert prenatal ultrasound measurements to Intergrowth z-scores/centiles (generic)
}
\examples{
# get centile for child at 100 gestational days with 11 cm head circumference
igpre_hccm2centile(100, 11)
}
\references{
International standards for fetal growth based on serial ultrasound measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project
Papageorghiou, Aris T et al.
The Lancet, Volume 384, Issue 9946, 869-879
}

