% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intergrowth_pre.R
\name{igpre_hccm2zscore}
\alias{igpre_accm2centile}
\alias{igpre_accm2zscore}
\alias{igpre_bpdcm2centile}
\alias{igpre_bpdcm2zscore}
\alias{igpre_flcm2centile}
\alias{igpre_flcm2zscore}
\alias{igpre_hccm2centile}
\alias{igpre_hccm2zscore}
\alias{igpre_ofdcm2centile}
\alias{igpre_ofdcm2zscore}
\title{Convert prenatal ultrasound measurements to Intergrowth z-scores/centiles}
\usage{
igpre_hccm2zscore(gagedays, hccm)

igpre_bpdcm2zscore(gagedays, bpdcm)

igpre_ofdcm2zscore(gagedays, ofdcm)

igpre_accm2zscore(gagedays, accm)

igpre_flcm2zscore(gagedays, flcm)

igpre_hccm2centile(gagedays, hccm)

igpre_bpdcm2centile(gagedays, bpdcm)

igpre_ofdcm2centile(gagedays, ofdcm)

igpre_accm2centile(gagedays, accm)

igpre_flcm2centile(gagedays, flcm)
}
\arguments{
\item{gagedays}{gestational age in days}

\item{hccm}{head circumference (cm) measurement(s) to convert}

\item{bpdcm}{biparietel diameter (cm) measurement(s) to convert}

\item{ofdcm}{occipito-frontal diameter (cm) measurement(s) to convert}

\item{accm}{abdominal circumference (cm) measurement(s) to convert}

\item{flcm}{femur length (cm) measurement(s) to convert}
}
\description{
Convert prenatal ultrasound measurements to Intergrowth z-scores/centiles
}
\examples{
# get centile for child at 100 gestational days with 11 cm head circumference
igpre_hccm2centile(100, 11)
}
\references{
International standards for fetal growth based on serial ultrasound measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project
Papageorghiou, Aris T et al.
The Lancet, Volume 384, Issue 9946, 869-879
}

