% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{panel.who}
\alias{panel.who}
\title{WHO growth standard panel function for lattice}
\usage{
panel.who(x, x_var = "agedays", y_var = "htcm", sex = "Female",
  quants = c(0.01, 0.05, 0.25, 0.5), alpha = 0.15, center = FALSE,
  labels = TRUE, x_trans = identity, y_trans = identity)
}
\arguments{
\item{x}{value or vector of values that correspond to a measure defined by \code{x_var}}

\item{x_var}{x variable name (typically "agedays") - see details}

\item{y_var}{y variable name (typically "htcm" or "wtkg") - see details}

\item{sex}{"Male" or "Female"}

\item{quants}{quantiles at which to draw the WHO polygons}

\item{alpha}{transparency of the polygons}

\item{center}{should the polygons be centered around the median?}

\item{labels}{should the quantiles be labeled? (not implemented)}

\item{x_trans}{transformation function to be applied to x-axis}

\item{y_trans}{transformation function to be applied to y-axis}
}
\description{
WHO growth standard panel function for lattice
}
\examples{
library(lattice)
xyplot(wtkg ~ agedays, data = subset(cpp, subjid == 8),
  panel = function(x, y, ...) {
    panel.who(x = seq(0, 2558, by = 30),
      sex = "Male", y_var = "wtkg", quants = pnorm(-3:0))
    panel.xyplot(x, y, ...)
  },
  col = "black"
)
}

