% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot.subjDiv}
\alias{plot.subjDiv}
\title{Plot a raw data trajectory}
\usage{
\method{plot}{subjDiv}(x, subjid, y_var = "htcm", center = FALSE,
  x_range = NULL, width = 500, height = 520, hover = NULL, p = 100 *
  pnorm(-3:0), x_units = c("days", "months", "years"), ...)
}
\arguments{
\item{x}{an object of class "subDiv" obtained from \code{\link{by_subject}}}

\item{subjid}{subject id of subject to plot}

\item{y_var}{name of y variable to model (usually an anthropometric measure or z-score scaled anthropometric measure)}

\item{center}{should the trajectory be centered around the median WHO standard?  This is equivalent to plotting the age difference score (like height-for-age difference - HAD)}

\item{x_range}{a vector specifying the range (min, max) that the superposed growth standard should span on the x-axis}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{hover}{variable names in \code{x$data} to show on hover for each point (only variables with non-NA data will be shown)}

\item{p}{centiles at which to draw the WHO polygons}

\item{x_units}{units of age x-axis (days, months, or years)}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a raw data trajectory
}
