% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_complete_pairs}
\alias{plot_complete_pairs}
\title{Plot a heat map of frequency of "complete" (both non-NA) pairs of variables}
\usage{
plot_complete_pairs(dat, subject = FALSE, width = 700, height = 700,
  thresh = 0.95, ...)
}
\arguments{
\item{dat}{data frame}

\item{subject}{should subject-level (TRUE) or time-varying (FALSE) variables be plotted?}

\item{width}{width of plot in pixels}

\item{height}{height of plot in pixels}

\item{thresh}{percentage NA threshold above which variables will be ignored (to help deal with cases involving many variables)}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a heat map of frequency of "complete" (both non-NA) pairs of variables
}
\details{
Subject-level variables are treated differently than time-varying variables in that they are repeated for each subject.  When plotting summaries of subject-level variables, the data is first subset to one record per subject.
}
\examples{
plot_complete_pairs(cpp)
plot_complete_pairs(cpp, subject = TRUE)
}
