% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_first_visit_age}
\alias{plot_first_visit_age}
\title{Plot histogram and quantile plot of age at first visit}
\usage{
plot_first_visit_age(dat, subjid = "subjid", agevar = "agedays",
  agelab = "first visit age (days)", width = 450, height = 450)
}
\arguments{
\item{dat}{a longitudinal growth study data set}

\item{subjid}{variable name in \code{dat} that contains the subject's identifier}

\item{agevar}{variable name in \code{dat} that contains a measure of the subject's age}

\item{agelab}{label of the age axis}

\item{width}{the width of each plot in pixels}

\item{height}{the height of each plot in pixels}
}
\description{
Plot histogram and quantile plot of age at first visit
}
\examples{
\donttest{
plot_first_visit_age(cpp)
}
}

