% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot a stacked bar chart indicating NAs for each variable in a dataset}
\usage{
plot_missing(dat, subject = FALSE, width = 800, height = 500, ...)
}
\arguments{
\item{dat}{data frame}

\item{subject}{should subject-level (TRUE) or time-varying (FALSE) variables be plotted?}

\item{width}{width of plot in pixels}

\item{height}{height of plot in pixels}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a stacked bar chart indicating NAs for each variable in a dataset
}
\details{
Subject-level variables are treated differently than time-varying variables in that they are repeated for each subject.  When plotting summaries of subject-level variables, the data is first subset to one record per subject.
}
\examples{
plot_missing(cpp)
plot_missing(cpp, subject = TRUE)
}

