% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/plot_multistudy.R
\name{plot_multi_subj_boxplot}
\alias{plot_multi_subj_boxplot}
\title{Plot boxplots of distrubutions of number of records per subject for a list of studies}
\usage{
plot_multi_subj_boxplot(dat_list, width = 800, height = 500)
}
\arguments{
\item{dat_list}{a list of data frames containing study data}

\item{width}{width of the plot in pixels}

\item{height}{height of the plot in pixels}
}
\description{
Plot boxplots of distrubutions of number of records per subject for a list of studies
}
\examples{
dat_list <- list(
  cpp1 = cpp,
  cpp2 = cpp,
  cpp3 = cpp
)
plot_multi_subj_boxplot(dat_list)
}

