% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multistudy.R
\name{plot_time_count_grid}
\alias{plot_time_count_grid}
\title{Plot counts by age for a list of studies}
\usage{
plot_time_count_grid(dat_list, agevar = "agedays",
  xlab = "Age since birth at examination (days)", width = 845,
  height = 80)
}
\arguments{
\item{dat_list}{a list of data frames containing study data}

\item{agevar}{variable name in \code{dat} that contains a measure of the subject's age}

\item{xlab}{label for x axis}

\item{width}{width of the plot in pixels}

\item{height}{height of each panel of the plot in pixels}
}
\description{
Plot counts by age for a list of studies
}
\examples{
dat_list <- list(
  cpp1 = cpp,
  cpp2 = cpp,
  cpp3 = cpp
)
plot_time_count_grid(dat_list)
}

