% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_univar}
\alias{plot_univar}
\title{Make a grid of univariate summary plots}
\usage{
plot_univar(dat, subject = FALSE, ncol = 3, width = 300, height = 300)
}
\arguments{
\item{dat}{data frame}

\item{subject}{should subject-level (TRUE) or time-varying (FALSE) variables be plotted?}

\item{ncol}{number of columns in the grid}

\item{width}{width of each plot in pixels}

\item{height}{height of each plot in pixels}
}
\description{
Make a grid of univariate summary plots
}
\details{
Subject-level variables are treated differently than time-varying variables in that they are repeated for each subject.  When plotting summaries of subject-level variables, the data is first subset to one record per subject.
}
\examples{
plot_univar(cpp, subject = TRUE)
plot_univar(cpp)
}

