% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multistudy.R
\name{plot_var_matrix}
\alias{plot_var_matrix}
\title{Plot a matrix comparing variables present in a list of studies}
\usage{
plot_var_matrix(dat_list, width = 845, h_padding = 0, head = NULL)
}
\arguments{
\item{dat_list}{a list of data frames containing study data}

\item{width}{width of the plot in pixels}

\item{h_padding}{extra height to add to the plot to account for long variable names}

\item{head}{the number of variables to limit the x-axis to (if negative, it will show all but the first \code{head} variables)}
}
\description{
Plot a matrix comparing variables present in a list of studies
}
\examples{
dat_list <- list(
  cpp1 = cpp[, c(1:5, 7:9, 14:19, 23:32)],
  cpp2 = cpp[, c(1:5, 11:24)],
  cpp3 = cpp
)
plot_var_matrix(dat_list)
}

