% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_visit_distn}
\alias{plot_visit_distn}
\title{Plot histogram and quantile plot of number of "visits" for each subject}
\usage{
plot_visit_distn(dat, width = 450, height = 450)
}
\arguments{
\item{dat}{a longitudinal growth study dataset}

\item{width}{the width of each plot in pixels}

\item{height}{the height of each plot in pixels}
}
\description{
Plot histogram and quantile plot of number of "visits" for each subject
}
\examples{
\donttest{
plot_visit_distn(cpp)
}
}

