% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_z}
\alias{plot_z}
\title{Plot a fitted trajectory on z-score scale}
\usage{
plot_z(x, x_range = NULL, nadir = FALSE, width = 500, height = 520,
  hover = NULL, checkpoints = TRUE, z = -3:0, x_units = c("days",
  "months", "years"), ...)
}
\arguments{
\item{x}{an object returned from \code{\link{fit_trajectory}}}

\item{x_range}{a vector specifying the range (min, max) that the superposed z-score bands should span on the x-axis}

\item{nadir}{should a guide be added to the plot showing the location of the nadir?}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{hover}{variable names in \code{x$data} to show on hover for each point (only variables with non-NA data will be shown)}

\item{checkpoints}{should the checkpoints be plotted (if available)?}

\item{z}{z-scores at which to draw the z-score bands}

\item{x_units}{units of age x-axis (days, months, or years)}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a fitted trajectory on z-score scale
}
\examples{
mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
plot_z(fit)
}

