% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_zvelocity}
\alias{plot_zvelocity}
\title{Plot a fitted trajectory's z-score velocity}
\usage{
plot_zvelocity(x, width = 500, height = 520, ...)
}
\arguments{
\item{x}{an object returned from \code{\link{fit_trajectory}}}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a fitted trajectory's z-score velocity
}
\examples{
fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
plot_zvelocity(fit)
}

