% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{SL.brokenstick}
\alias{SL.brokenstick}
\title{Super Learner wrapper for the brokenstick method}
\usage{
SL.brokenstick(Y, X, newX = X, family = gaussian(), knots = 6,
  mn = NULL, mx = NULL, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{\ldots}{additional parameters passed to \code{\link[brokenstick]{brokenstick}}, most notably \code{knots}}
}
\description{
Super Learner wrapper for the brokenstick method
}
\examples{
\dontrun{
smc <- get_smocc_data()[1:2000,]
bsfit <- get_fit(smc, y_var = "haz", method = "brokenstick")
fit <- fit_trajectory(subset(smc, subjid == 10001), fit = bsfit)
plot(fit)
}
}

