% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{SL.sitar}
\alias{SL.sitar}
\title{Get the result of fitting sitar to a data set}
\usage{
SL.sitar(Y, X, newX = X, family = gaussian(), df = 3, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{\ldots}{additional parameters passed to \code{\link[sitar]{sitar}}, most notably \code{df} which defaults to 3}
}
\description{
Get the result of fitting sitar to a data set
}
\details{
This essentially gets an anthropometric data set into shape for \code{\link[sitar]{sitar}} (sets appropriate data structure and removes missing values) and runs the fitting routine.
}
\note{
The settings for \code{x_trans} and \code{y_trans} must match that used in \code{\link{fit_trajectory}} and appropriate inverse transformations must be set there accordingly as well.
}
\examples{
\dontrun{
smc <- get_smocc_data()[1:2000,]
sitfit <- get_fit(smc, y_var = "htcm", method = "sitar")
fit <- fit_trajectory(subset(smc, subjid == 10001), fit = sitfit)
plot(fit)
}
}

