% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{trscope_trajectories}
\alias{trscope_trajectories}
\title{Create Trelliscope display of subject trajectories}
\usage{
trscope_trajectories(dat, center = FALSE, who_range = NULL, width = 500,
  height = 520, name = NULL, desc = "", group = NULL,
  vdbConn = getOption("vdbConn"), view = TRUE)
}
\arguments{
\item{dat}{either a data frame or object created by \code{\link{by_subject}} or \code{\link{fit_all_trajectories}}}

\item{center}{should the trajectory be centered around the median WHO standard?}

\item{who_range}{a vector specifying the range (min, max) that the superposed WHO growth standard should span on the x-axis}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{name}{name of the Trelliscope display (if left NULL, will be implied by the vairables being plotted)}

\item{desc}{description of the Trelliscope display}

\item{group}{group in which to place the Trelliscope display}

\item{vdbConn}{an optional VDB connection}

\item{view}{should the plot be viewed upon creation?}
}
\description{
Create Trelliscope display of subject trajectories
}
\examples{
\donttest{
  cppsubj <- by_subject(cpp)
  cppt <- fit_all_trajectories(cppsubj, method = "rlm")
  cppt <- trscope_trajectories(cppt)
}
}

