% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{trscope_velocities}
\alias{trscope_velocities}
\title{Create Trelliscope display of the velocities of per-subject fitted trajectories}
\usage{
trscope_velocities(dat, z = FALSE, x_range = NULL, width = 500,
  height = 520, name = NULL, desc = "", group = NULL,
  vdbConn = getOption("vdbConn"))
}
\arguments{
\item{dat}{either a data frame or object created by \code{\link{by_subject}} or \code{\link{fit_all_trajectories}}}

\item{z}{should velocities according to z-score scale or raw scale be plotted?}

\item{x_range}{a vector specifying the range (min, max) that the x-axis should span}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{name}{name of the Trelliscope display (if NULL, will be implied by the variables being plotted and the method name)}

\item{desc}{description of the Trelliscope display}

\item{group}{group in which to place the Trelliscope display}

\item{vdbConn}{an optional VDB connection}
}
\description{
Create Trelliscope display of the velocities of per-subject fitted trajectories
}
\examples{
\donttest{
  cppsubj <- by_subject(cpp)
  cppt <- fit_all_trajectories(cppsubj, method = "rlm")
  cppt <- trscope_velocities(cppt)
}
}

