% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisions.R
\name{by_trajectory_checkpoints}
\alias{by_trajectory_checkpoints}
\title{Split by-subject trajectory-fitted data by checkpoint categorizations}
\usage{
by_trajectory_checkpoints(dat, complete = TRUE)
}
\arguments{
\item{dat}{a data object returned by \code{\link{fit_all_trajectories}}}

\item{complete}{subset only to those that have fitted checkpoints}
}
\description{
Split by-subject trajectory-fitted data by checkpoint categorizations
}
\examples{
\donttest{
cppsubj <- by_subject(cpp)
cppt <- fit_all_trajectories(cppsubj, method = "rlm")
cppcp <- by_trajectory_checkpoints(cppt)
}
}

