% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.face}
\alias{fit_method.face}
\title{Get the result of fitting face.sparse to a dataset}
\usage{
\method{fit_method}{face}(dat, ...)
}
\arguments{
\item{dat}{data frame containing variables to model}

\item{\ldots}{additional parameters passed to \code{\link[face]{face.sparse}}, most notably \code{knots} which defaults to 10}
}
\description{
Get the result of fitting face.sparse to a dataset
}
\details{
This essentially gets an anthropometric dataset into shape for \code{\link[face]{face.sparse}} (sets appropriate data structure and removes missing values) and runs the fitting routine.
}
\note{
The settings for \code{x_trans} and \code{y_trans} must match that used in \code{\link{fit_trajectory}} and appropriate inverse transformations must be set there accordingly as well.
}
\examples{
\dontrun{
facefit <- get_fit(cpp, y_var = "haz", method = "face")
fit <- fit_trajectory(subset(cpp, subjid == 2), fit = facefit)
plot(fit)
}
}

