% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/analysis_fit_methods.R
\name{fit_method.fda}
\alias{fit_method.fda}
\title{Compute functional "fda" fit of growth trajectory}
\usage{
\method{fit_method}{fda}(dat, ...)
}
\arguments{
\item{dat}{data frame specifying x and y coordinates of data to fit}

\item{\ldots}{additional parameters passed to \code{\link[fda]{smooth.basisPar}}, notably \code{lambda} which defaults to 0.1}
}
\description{
Compute functional "fda" fit of growth trajectory
}
\note{
The trajectory fitting functions are most easily accessed through calling \code{\link{fit_trajectory}} with the \code{method} argument to specify the modeling approach to use.

These fitting functions can easily be replaced by simply calling the associated R methods, but are provided for convenience to standardize input/output to simplify swapping fitting methods.
}

