% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fix_height}
\alias{fix_height}
\title{Merge 'htcm' and 'lencm' into the 'htcm' variable}
\usage{
fix_height(dat)
}
\arguments{
\item{dat}{data}
}
\description{
Merge 'htcm' and 'lencm' into the 'htcm' variable
}

