% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{get_agefreq}
\alias{get_agefreq}
\title{Get age frequency}
\usage{
get_agefreq(dat, age_range = NULL)
}
\arguments{
\item{dat}{a longitudinal growth study dataset}

\item{age_range}{optional range to ....}
}
\description{
Get age frequency
}
\examples{
\donttest{
agefreq <- get_agefreq(cpp)
plot_agefreq(agefreq)
}
}

