% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{get_nadir}
\alias{get_nadir}
\title{Get nadir of a growth velocity}
\usage{
get_nadir(obj)
}
\arguments{
\item{obj}{object created from \code{\link{fit_trajectory}}}
}
\description{
Get nadir of a growth velocity
}

