% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{grid_deriv}
\alias{grid_deriv}
\title{Estimate derivative given a grid of points}
\usage{
grid_deriv(x, y)
}
\arguments{
\item{x}{x variable (should be a regularly-spaced grid of points)}

\item{y}{y variable}
}
\description{
Estimate derivative given a grid of points
}

