% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbgd-package.R
\docType{package}
\name{hbgd-package}
\alias{hbgd}
\alias{hbgd-package}
\title{hbgd: Healthy Birth, Growth & Development}
\description{
A package for visual and analytical methods for the analysis of
longitudinal growth data.
}
\details{
\url{http://hbgdki.github.io/hbgd/}
}
\examples{
help(package = "hbgd")
}

