% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot.fittedTrajectory}
\alias{plot.fittedTrajectory}
\title{Plot a fitted trajectory}
\usage{
\method{plot}{fittedTrajectory}(x, center = FALSE, x_range = NULL,
  width = 500, height = 520, hover = NULL, checkpoints = TRUE, p = 100
  * pnorm(-3:0), ...)
}
\arguments{
\item{x}{an object returned from \code{\link{fit_trajectory}}}

\item{center}{should the trajectory be centered around the median WHO standard?  This is equivalent to plotting the age difference score (like height-for-age difference - HAD)}

\item{x_range}{a vector specifying the range (min, max) that the superposed growth standard should span on the x-axis}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{hover}{variable names in \code{x$data} to show on hover for each point (only variables with non-NA data will be shown)}

\item{checkpoints}{should the checkpoints be plotted (if available)?}

\item{p}{centiles at which to draw the WHO polygons}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a fitted trajectory
}
\examples{
fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
plot(fit)
plot(fit, center = TRUE)
plot(fit, hover = c("wtkg", "bmi", "waz", "haz"))
}

