% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_agefreq}
\alias{plot_agefreq}
\title{Plot age frequency}
\usage{
plot_agefreq(agefreq, xlab = "Age since birth at examination (days)",
  ylab = "# examinations", width = 700, height = 350)
}
\arguments{
\item{agefreq}{an object returned from \code{\link{get_agefreq}}}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{width}{width of plot in pixels}

\item{height}{height of plot in pixels}
}
\description{
Plot age frequency
}
\examples{
\donttest{
agefreq <- get_agefreq(cpp)
plot_agefreq(agefreq)
}
}

