% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_first_visit_age}
\alias{plot_first_visit_age}
\title{Plot histogram and quantile plot of age at first visit}
\usage{
plot_first_visit_age(dat, agelab = "first visit age (days)", width = 450,
  height = 450)
}
\arguments{
\item{dat}{a longitudinal growth study dataset}

\item{agelab}{label of the age axis}

\item{width}{the width of each plot in pixels}

\item{height}{the height of each plot in pixels}
}
\description{
Plot histogram and quantile plot of age at first visit
}
\examples{
\donttest{
plot_first_visit_age(cpp)
}
}

