% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_zvelocity}
\alias{plot_zvelocity}
\title{Plot a fitted trajectory's z-score velocity}
\usage{
plot_zvelocity(x, width = 500, height = 520, x_units = c("days", "months",
  "years"), ...)
}
\arguments{
\item{x}{an object returned from \code{\link{fit_trajectory}}}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{x_units}{units of age x-axis (days, months, or years)}

\item{\ldots}{additional parameters passed to \code{\link{figure}}}
}
\description{
Plot a fitted trajectory's z-score velocity
}
\examples{
mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
plot_zvelocity(fit)
}

