% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trelliscope.R
\name{trscope_velocities}
\alias{trscope_velocities}
\title{Create Trelliscope display of the velocities of per-subject fitted trajectories}
\usage{
trscope_velocities(dat, z = FALSE, x_range = NULL, width = 500,
  height = 520, name = NULL, desc = "", group = NULL,
  vdb_conn = getOption("vdbConn"), nadir = FALSE, recovery = NULL,
  x_units = "days")
}
\arguments{
\item{dat}{either a data frame or object created by \code{\link{by_subject}} or \code{\link{fit_all_trajectories}}}

\item{z}{should velocities according to z-score scale or raw scale be plotted?}

\item{x_range}{a vector specifying the range (min, max) that the x-axis should span}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{name}{name of the Trelliscope display (if NULL, will be implied by the variables being plotted and the method name)}

\item{desc}{description of the Trelliscope display}

\item{group}{group in which to place the Trelliscope display}

\item{vdb_conn}{an optional VDB connection}

\item{nadir}{should a guide be added to the plot showing the location of the nadir? (only valid when z = TRUE)}

\item{recovery}{age in days at which to plot recovery from nadir (only valid when z = TRUE) - if NULL (default), will not be plotted}

\item{x_units}{units of age x-axis (days, months, or years)}
}
\description{
Create Trelliscope display of the velocities of per-subject fitted trajectories
}
\examples{
\dontrun{
cppsubj <- by_subject(cpp)
cppfit  <- get_fit(cpp, method = "rlm")
cpptr   <- fit_all_trajectories(cppsubj, cppfit)
cppplot <- trscope_velocities(cpptr)
}
}

